createname sayfası component yapısına ayrıldı
