+++ app/(pages)/(auth)/createname/title-section.tsx
@@ -1,0 +1,15 @@
+const TitleSection = () => {
+  return (
+    <>
+      <h2 className="font-bold text-xl text-white">
+        Realtime Chat Projesine Hoş Geldiniz
+      </h2>
+      <p className="text-white text-sm max-w-sm mt-2">
+        Giriş yapabilmen için son bir adım kaldı. Kendine herkesten farklı bir
+        kullanıcı adı oluşturmalısın.
+      </p>
+    </>
+  );
+};
+
+export default TitleSection;
