+++ package.json
@@ -19,6 +19,7 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "framer-motion": "^11.2.10",
+    "jwt-decode": "^4.0.0",
     "lucide-react": "^0.376.0",
     "mini-svg-data-uri": "^1.4.4",
     "next": "14.2.3",
